Fixed Makefile
When cloned an obj directory is supposed to be present. To fix this I
Added a '.stub' folder so that the directory gets added.

I also modded the makefile to include files with the -I flag
to automatically add ./include to the root rather than using a relative directory
So now includes go from this:
\#include '../include/foobar.h'
to this:
\#include <foobar.h>

Its a minor change but it needs to be done.
I'm completely abandoning this project.
I've totally lost interest but I just need to make sure it can be built by
people who don't know C, which is obviously a challenge as the build tools
for this programming language are ancient and nobody uses it anymore.